+++ sqlite3_web/lib/src/client.dart
@@ -20,7 +20,7 @@ final class _CommitOrRollbackStream {
 }
 
 final class RemoteDatabase implements Database {
-  final WorkerConnection connection;
+  final FinalizableWorkerConnection finalizableConnection;
   final int databaseId;
   final Completer<Object?> _isClosed = Completer();
 
@@ -30,7 +30,12 @@ final class RemoteDatabase implements Database {
   final _CommitOrRollbackStream _commits = _CommitOrRollbackStream();
   final _CommitOrRollbackStream _rollbacks = _CommitOrRollbackStream();
 
-  RemoteDatabase({required this.connection, required this.databaseId}) {
+  WorkerConnection get connection => finalizableConnection.connection;
+
+  RemoteDatabase({
+    required this.finalizableConnection,
+    required this.databaseId,
+  }) {
     connection.closed.then((_) {
       if (!isClosed) {
         _isClosed.complete();
@@ -368,6 +373,24 @@ final class WorkerConnection extends ProtocolChannel {
     );
   }
 
+  @override
+  void handleNotification(Notification notification) {
+    notifications.add(notification);
+  }
+}
+
+final class FinalizableWorkerConnection {
+  final WorkerConnection connection;
+
+  FinalizableWorkerConnection(this.connection) {
+    ProtocolChannel.finalizer.attach(this, connection, detach: this);
+  }
+
+  void close() {
+    connection.close();
+    ProtocolChannel.finalizer.detach(this);
+  }
+
   Future<RemoteDatabase> requestDatabase({
     required String wasmUri,
     required String databaseName,
@@ -375,7 +398,7 @@ final class WorkerConnection extends ProtocolChannel {
     required bool onlyOpenVfs,
     required JSAny? additionalOptions,
   }) async {
-    final response = await sendRequest(
+    final response = await connection.sendRequest(
       newOpenRequest(
         requestId: 0,
         wasmUri: wasmUri,
@@ -388,14 +411,9 @@ final class WorkerConnection extends ProtocolChannel {
     );
 
     return RemoteDatabase(
-      connection: this,
+      finalizableConnection: this,
       databaseId: (response.response as JSNumber).toDartInt,
     );
-  }
-
-  @override
-  void handleNotification(Notification notification) {
-    notifications.add(notification);
   }
 }
 
@@ -407,11 +425,11 @@ final class DatabaseClient implements WebSqlite {
 
   final Mutex _startWorkers = Mutex();
   bool _startedWorkers = false;
-  WorkerConnection? _connectionToDedicated;
-  WorkerConnection? _connectionToShared;
-  WorkerConnection? _connectionToDedicatedInShared;
+  FinalizableWorkerConnection? _connectionToDedicated;
+  FinalizableWorkerConnection? _connectionToShared;
+  FinalizableWorkerConnection? _connectionToDedicatedInShared;
 
-  WorkerConnection? _connectionToLocal;
+  FinalizableWorkerConnection? _connectionToLocal;
 
   final Set<MissingBrowserFeature> _missingFeatures = {};
 
@@ -438,8 +456,10 @@ final class DatabaseClient implements WebSqlite {
     });
   }
 
-  WorkerConnection _connection(ConnectableChannel channel) {
-    return WorkerConnection(channel, _handleCustomRequest);
+  FinalizableWorkerConnection _connection(ConnectableChannel channel) {
+    return FinalizableWorkerConnection(
+      WorkerConnection(channel, _handleCustomRequest),
+    );
   }
 
   Future<void> _startDedicated() async {
@@ -491,14 +511,14 @@ final class DatabaseClient implements WebSqlite {
     _connectionToShared = _connection(channel);
   }
 
-  Future<WorkerConnection> _connectToDedicatedInShared() {
+  Future<FinalizableWorkerConnection> _connectToDedicatedInShared() {
     return _startWorkers.withCriticalSection(() async {
       if (_connectionToDedicatedInShared case final conn?) {
         return conn;
       }
 
       final (endpoint, channel) = await createChannel();
-      await _connectionToShared!.sendRequest(
+      await _connectionToShared!.connection.sendRequest(
         newConnectRequest(requestId: 0, endpoint: endpoint, databaseId: null),
         MessageType.simpleSuccessResponse,
       );
@@ -507,7 +527,7 @@ final class DatabaseClient implements WebSqlite {
     });
   }
 
-  Future<WorkerConnection> _connectToLocal() async {
+  Future<FinalizableWorkerConnection> _connectToLocal() async {
     return _startWorkers.withCriticalSection(() async {
       if (_connectionToLocal case final conn?) {
         return conn;
@@ -647,10 +667,10 @@ final class DatabaseClient implements WebSqlite {
     }
 
     if (_connectionToDedicated case final dedicated?) {
-      await dedicatedCompatibilityCheck(dedicated);
+      await dedicatedCompatibilityCheck(dedicated.connection);
     }
     if (_connectionToShared case final shared?) {
-      await sharedCompatibilityCheck(shared);
+      await sharedCompatibilityCheck(shared.connection);
     }
 
     available.add(DatabaseImplementation.inMemoryLocal);
@@ -672,7 +692,7 @@ final class DatabaseClient implements WebSqlite {
     );
 
     return RemoteDatabase(
-      connection: channel,
+      finalizableConnection: channel,
       // The database id for this pre-existing connection is always zero.
       // It gets assigned by the worker handling the OpenAdditonalConnection
       // request.
@@ -689,7 +709,7 @@ final class DatabaseClient implements WebSqlite {
   }) async {
     await startWorkers();
 
-    WorkerConnection connection;
+    FinalizableWorkerConnection connection;
     switch (implementation.access) {
       case AccessMode.throughSharedWorker:
         if (implementation.storage == StorageMode.opfs) {
@@ -758,7 +778,12 @@ final class DatabaseClient implements WebSqlite {
   }
 
   @override
-  void close() {}
+  void close() {
+    _connectionToShared?.close();
+    _connectionToDedicatedInShared?.close();
+    _connectionToDedicated?.close();
+    _connectionToLocal?.close();
+  }
 
   /// Compares available ways to access databases by the performance and
   /// and reliability of the implementation.