+++ sqlite3_web/lib/src/database.dart
@@ -335,7 +335,11 @@ abstract class WebSqlite {
     JSAny? additionalOptions,
   });
 
-  /// Closes this instance and associated workers.
+  /// Closes this instance and associated dedicated workers.
+  ///
+  /// This package will register a [Finalizer] closing workers automatically, so
+  /// calling this method is not required. However, it can still be used to be
+  /// explicit about when to close resources.
   void close();
 
   /// Entrypoints for workers hosting databases.